% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aricode.R
\name{Chi2}
\alias{Chi2}
\title{Chi-square statistics}
\usage{
Chi2(c1, c2)
}
\arguments{
\item{c1}{a vector of length n with value between 0 and N1 < n containing the first classification.
Must be a vector of integers or characters, a numeric, a factor, but not a list. Avoid character.}

\item{c2}{a vector of length n with value between 0 and N2 < n containing the second classification.
Must be a vector of integers or characters, a numeric or a factor, but not a list. Avoid character.}
}
\value{
a scalar with the chi-square statistics.
}
\description{
A function to compute the Chi-2 statistics
}
\examples{
data(iris)
cl <- cutree(hclust(dist(iris[, -5])), 4)
Chi2(cl, iris$Species)
}
\seealso{
\code{\link{ARI}}, \code{\link{NID}}, \code{\link{NVI}}, \code{\link{NMI}}, \code{\link{clustComp}}
}
