% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aricode.R
\name{clustComp}
\alias{clustComp}
\title{Measures of similarity between two classification}
\usage{
clustComp(c1, c2, sorted_pairs = NULL, AMI = FALSE)
}
\description{
`r lifecycle::badge("deprecated")`

A function for computing all the measures of similarity implemented in this package at once.
Include ARI, RI, MI VI, NVI, ID, NID, NMI, Chi2, MARI, Frobenius

Just a change in the function name: please use [compare_clustering()].
}
\examples{
data(iris)
cl <- cutree(hclust(dist(iris[, -5])), 4)
out <- sort_pairs(cl, iris$Species)
# ->
out <- sort_pairs(cl, iris$Species)
}
\keyword{internal}
