% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcul_Rand_CtoR.R
\name{clustCompM}
\alias{clustCompM}
\title{measures of similarity between classification}
\usage{
clustCompM(c1, c2)
}
\description{
A function various measures of similarity between two classifications
}

