% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aricode.R
\name{compare_clustering}
\alias{compare_clustering}
\title{Measures of similarity between two classification}
\usage{
compare_clustering(c1, c2, sorted_pairs = NULL, AMI = FALSE)
}
\arguments{
\item{c1}{A vector of length $n$ with values between 0 and $N_1 < n$ representing the
first classification. Supported types: integer, numeric, or factor. Avoid character
vectors for better performance. Must not be a list.}

\item{c2}{A vector of length $n$ with values between 0 and $N_2 < n$ representing the
second classification. Supported types: integer, numeric, or factor. Avoid character
vectors for better performance. Must not be a list.}

\item{sorted_pairs}{optional output of function sort_pairs (if already computed).
If `NULL` (the default), will be called internally}

\item{AMI}{Boolean: should the AMI be computed (more costly than all other measures)? Default is `FALSE`.}
}
\value{
a list with all the measures available
}
\description{
A function for computing all the measures of similarity implemented in this package at once.
Include (A)RI, (N)MI, (N)VI, (N)ID, Chi2, MARI, Frobenius
}
\examples{
data(iris)
cl <- cutree(hclust(dist(iris[, -5])), 4)
compare_clustering(cl, iris$Species)
}
\seealso{
\code{\link{RI}}, \code{\link{NID}}, \code{\link{NVI}}, \code{\link{NMI}}, \code{\link{ARI}}
}
