% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aricode.R
\name{entropy}
\alias{entropy}
\title{Entropy}
\usage{
entropy(c1, c2)
}
\arguments{
\item{c1}{a vector of length n with value between 0 and N1 < n containing the first classification.
Must be a vector of integers or characters, a numeric, a factor, but not a list. Avoid character.}

\item{c2}{a vector of length n with value between 0 and N2 < n containing the second classification.
Must be a vector of integers or characters, a numeric or a factor, but not a list. Avoid character.}
}
\value{
a list with the two conditional entropies, the joint entropy and output of sortPairs.
}
\description{
A function to compute the empirical entropy for two vectors of classification and the joint entropy
}
\examples{
data(iris)
cl <- cutree(hclust(dist(iris[, -5])), 4)
entropy(cl, iris$Species)
}
