% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aricode.R
\name{Frobenius}
\alias{Frobenius}
\title{Frobenius norm}
\usage{
Frobenius(c1, c2)
}
\arguments{
\item{c1}{a vector of length n with value between 0 and N1 < n containing the first classification.
Must be a vector of integers or characters, a numeric, a factor, but not a list. Avoid character.}

\item{c2}{a vector of length n with value between 0 and N2 < n containing the second classification.
Must be a vector of integers or characters, a numeric or a factor, but not a list. Avoid character.}
}
\value{
a scalar with the chi-square statistics.
}
\description{
A function to compute the Frobenius norm between two classification as defined in Lajugie et al. 2014 and Arlot et al 2019
}
\examples{
data(iris)
cl <- cutree(hclust(dist(iris[, -5])), 4)
Frobenius(cl, iris$Species)
}
\references{
- Rémi Lajugie, Francis Bach, and Sylvain Arlot. "Large-margin metric learning for constrained partitioning problems." International Conference on Machine Learning. PMLR, 2014.
  - Sylvain Arlot , Alain Celisse, and Zaid Harchaoui. "A kernel multiple change-point algorithm via model selection." Journal of machine learning research 20.162 (2019): 1-56.
}
\seealso{
\code{\link{ARI}}, \code{\link{NID}}, \code{\link{NVI}}, \code{\link{NMI}}, \code{\link{clustComp}}
}
