% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aricode.R
\name{MARIraw}
\alias{MARIraw}
\title{raw Modified Adjusted Rand Index}
\usage{
MARIraw(c1, c2, sorted_pairs = NULL)
}
\description{
`r lifecycle::badge("deprecated")`

The function MARI now owns an argument function `raw` if one wishes to compute the raw version of MARI.
}
\details{
A function to compute a modified adjusted rand index between two classifications as proposed by Sundqvist (2023), based on a multinomial model.
Raw means that the index is not divided by the (maximum - expected) value.
}
\examples{
data(iris)
cl <- cutree(hclust(dist(iris[, -5])), 4)
out <- MARIraw(cl, iris$Species)
# ->
out <- MARI(cl, iris$Species, raw = TRUE)
}
\keyword{internal}
