% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aricode.R
\name{NVI}
\alias{NVI}
\title{Normalized variation of information (NVI)}
\usage{
NVI(c1, c2, sorted_pairs = NULL)
}
\arguments{
\item{c1}{A vector of length $n$ with values between 0 and $N_1 < n$ representing the
first classification. Supported types: integer, numeric, or factor. Avoid character
vectors for better performance. Must not be a list.}

\item{c2}{A vector of length $n$ with values between 0 and $N_2 < n$ representing the
second classification. Supported types: integer, numeric, or factor. Avoid character
vectors for better performance. Must not be a list.}

\item{sorted_pairs}{optional output of function sort_pairs (if already computed).
If `NULL` (the default), will be called internally}
}
\value{
a scalar with the normalized variation of information.
}
\description{
A function to compute the NVI between two classifications
}
\examples{
data(iris)
cl <- cutree(hclust(dist(iris[, -5])), 4)
NVI(cl, iris$Species)
}
\seealso{
\code{\link{RI}}, \code{\link{NID}}, \code{\link{NMI}}, \code{\link{ARI}}, \code{\link{clustComp}}
}
