% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aricode.R
\name{sort_pairs}
\alias{sort_pairs}
\title{Sort Pairs}
\usage{
sort_pairs(c1, c2, spMat = FALSE)
}
\arguments{
\item{c1}{A vector of length $n$ with values between 0 and $N_1 < n$ representing the
first classification. Supported types: integer, numeric, or factor. Avoid character
vectors for better performance. Must not be a list.}

\item{c2}{A vector of length $n$ with values between 0 and $N_2 < n$ representing the
second classification. Supported types: integer, numeric, or factor. Avoid character
vectors for better performance. Must not be a list.}

\item{spMat}{Logical. If \code{TRUE}, returns the contingency table as a
sparse matrix. Note: sparse encoding may be more computationally expensive
than the algorithm itself. Default is \code{FALSE}.}
}
\value{
A list containing the following elements:
\itemize{
  \item \strong{spMat}: A sparsely encoded contingency matrix (only if \code{spMat = TRUE}).
  \item \strong{levels}: A list containing the retained levels for each classification.
  \item \strong{nij}: A vector of positive pair counts.
  \item \strong{ni., n.j}: Vectors of class counts for \code{c1} and \code{c2}, respectively.
  \item \strong{pair_c1, pair_c2}: Integer vectors specifying the classes in \code{c1} and \code{c2}
  corresponding to the counts in \code{nij}. These provide the row and column indices for
  the contingency matrix.
}
}
\description{
A function to sort pairs of integers or factors and identify the pairs between two classifications
}
\details{
Pair sorting, which is at the heart of computing all clustering comparison measures, has been carefully
optimized. Hence, even basic R operations (checking for the presence of NAs, type conversion, or
constructing a sparse contingency matrix as an output) have non-negligible cost compared to the pair
sorting itself. For optimal performance, please provide the vectors as integers or factors without any NAs.
}
\examples{
data(iris)
cl <- cutree(hclust(dist(iris[, -5])), 4)
out <- sort_pairs(cl, iris$Species)

}
