% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aricode.R
\name{AMI}
\alias{AMI}
\title{Adjusted Mutual Information}
\usage{
AMI(c1, c2)
}
\arguments{
\item{c1}{a vector containing the labels of the first classification. Must be a vector of characters, integers, numerics, or a factor, but not a list.}

\item{c2}{a vector containing the labels of the second classification.}
}
\value{
a scalar with the adjusted rand index.
}
\description{
A function to compute the adjusted mutual information between two classifications
}
\examples{
data(iris)
cl <- cutree(hclust(dist(iris[,-5])), 4)
AMI(cl,iris$Species)
}
\seealso{
\code{\link{ARI}}, \code{\link{RI}}, \code{\link{NID}}, \code{\link{NVI}}, \code{\link{NMI}}, \code{\link{clustComp}}
}
