% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aricode.R
\name{ARI}
\alias{ARI}
\title{Adjusted Rand Index}
\usage{
ARI(c1, c2)
}
\arguments{
\item{c1}{a vector of length n with value between 0 and N1 < n containing the first classification.
Must be a vector of integers or characters, a numeric, a factor, but not a list. Avoid character.}

\item{c2}{a vector of length n with value between 0 and N2 < n containing the second classification.
Must be a vector of integers or characters, a numeric or a factor, but not a list. Avoid character.}
}
\value{
a scalar with the adjusted Rand index.
}
\description{
A function to compute the adjusted rand index between two classifications
}
\examples{
data(iris)
cl <- cutree(hclust(dist(iris[, -5])), 4)
ARI(cl, iris$Species)
}
\seealso{
\code{\link{RI}}, \code{\link{NID}}, \code{\link{NVI}}, \code{\link{NMI}}, \code{\link{clustComp}}
}
