% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aricode.R
\name{Chi2}
\alias{Chi2}
\title{Chi-square statistics}
\usage{
Chi2(c1, c2, sorted_pairs = NULL)
}
\arguments{
\item{c1}{A vector of length $n$ with values between 0 and $N_1 < n$ representing the
first classification. Supported types: integer, numeric, or factor. Avoid character
vectors for better performance. Must not be a list.}

\item{c2}{A vector of length $n$ with values between 0 and $N_2 < n$ representing the
second classification. Supported types: integer, numeric, or factor. Avoid character
vectors for better performance. Must not be a list.}

\item{sorted_pairs}{optional output of function sort_pairs (if already computed).
If `NULL` (the default), will be called internally}
}
\value{
a scalar with the chi-square statistics.
}
\description{
A function to compute the Chi-2 statistics
}
\examples{
data(iris)
cl <- cutree(hclust(dist(iris[, -5])), 4)
Chi2(cl, iris$Species)
}
\seealso{
\code{\link{ARI}}, \code{\link{NID}}, \code{\link{NVI}}, \code{\link{NMI}}, \code{\link{clustComp}}
}
