% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aricode.R
\name{CHI2}
\alias{CHI2}
\title{Chi-square statistics}
\usage{
CHI2(c1, c2)
}
\arguments{
\item{c1}{a vector containing the labels of the first classification. Must be a vector of characters, integers, numerics, or a factor, but not a list.}

\item{c2}{a vector containing the labels of the second classification.}
}
\value{
a scalar with the rand index.
}
\description{
A function to compute the Chi-2 statistics
}
\examples{
data(iris)
cl <- cutree(hclust(dist(iris[,-5])), 4)
CHI2(cl,iris$Species)
}
\seealso{
\code{\link{ARI}}, \code{\link{NID}}, \code{\link{NVI}}, \code{\link{NMI}}, \code{\link{clustComp}}
}
