% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aricode.R
\name{Frobenius}
\alias{Frobenius}
\title{Frobenius norm}
\usage{
Frobenius(c1, c2, sorted_pairs = NULL)
}
\arguments{
\item{c1}{A vector of length $n$ with values between 0 and $N_1 < n$ representing the
first classification. Supported types: integer, numeric, or factor. Avoid character
vectors for better performance. Must not be a list.}

\item{c2}{A vector of length $n$ with values between 0 and $N_2 < n$ representing the
second classification. Supported types: integer, numeric, or factor. Avoid character
vectors for better performance. Must not be a list.}

\item{sorted_pairs}{optional output of function sort_pairs (if already computed).
If `NULL` (the default), will be called internally}
}
\value{
a scalar with the Frobenius norm.
}
\description{
A function to compute the Frobenius norm between two classifications as defined in Lajugie et al. 2014 and Arlot et al 2019
}
\examples{
data(iris)
cl <- cutree(hclust(dist(iris[, -5])), 4)
Frobenius(cl, iris$Species)
}
\references{
- Rémi Lajugie, Francis Bach, and Sylvain Arlot. "Large-margin metric learning for constrained partitioning problems." International Conference on Machine Learning. PMLR, 2014.
  - Sylvain Arlot , Alain Celisse, and Zaid Harchaoui. "A kernel multiple change-point algorithm via model selection." Journal of machine learning research 20.162 (2019): 1-56.
}
\seealso{
\code{\link{ARI}}, \code{\link{NID}}, \code{\link{NVI}}, \code{\link{NMI}}, \code{\link{clustComp}}
}
