% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aricode.R
\name{MARIraw}
\alias{MARIraw}
\title{raw Modified Adjusted Rand Index}
\usage{
MARIraw(c1, c2)
}
\arguments{
\item{c1}{a vector of length n with value between 0 and N1 < n containing the first classification.
Must be a vector of integers or characters, a numeric, a factor, but not a list. Avoid character.}

\item{c2}{a vector of length n with value between 0 and N2 < n containing the second classification.
Must be a vector of integers or characters, a numeric or a factor, but not a list. Avoid character.}
}
\value{
a scalar with the modified ARI without the division by the (maximum - expected)
}
\description{
A function to compute a modified adjusted rand index between two classifications as proposed by Sundqvist et al. in prep, based on a multinomial model. Raw means, that the index is not divided by the (maximum - expected) value.
}
\examples{
data(iris)
cl <- cutree(hclust(dist(iris[, -5])), 4)
MARIraw(cl, iris$Species)
}
\seealso{
\code{\link{ARI}}, \code{\link{NID}}, \code{\link{NVI}}, \code{\link{NMI}}, \code{\link{clustComp}}
}
