% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aricode.R
\name{NMI}
\alias{NMI}
\title{Normalized mutual information (NMI)}
\usage{
NMI(
  c1,
  c2,
  variant = c("max", "min", "sqrt", "sum", "joint"),
  sorted_pairs = NULL
)
}
\arguments{
\item{c1}{A vector of length $n$ with values between 0 and $N_1 < n$ representing the
first classification. Supported types: integer, numeric, or factor. Avoid character
vectors for better performance. Must not be a list.}

\item{c2}{A vector of length $n$ with values between 0 and $N_2 < n$ representing the
second classification. Supported types: integer, numeric, or factor. Avoid character
vectors for better performance. Must not be a list.}

\item{variant}{a string in ("max", "min", "sqrt", "sum", "joint"): different variants of NMI. Default use "max".}

\item{sorted_pairs}{optional output of function sort_pairs (if already computed).
If `NULL` (the default), will be called internally}
}
\value{
a scalar with the normalized mutual information .
}
\description{
A function to compute the NMI between two classifications
}
\examples{
data(iris)
cl <- cutree(hclust(dist(iris[, -5])), 4)
NMI(cl, iris$Species)
}
\seealso{
\code{\link{RI}}, \code{\link{NID}}, \code{\link{NVI}}, \code{\link{ARI}}, \code{\link{clustComp}}
}
