% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aricode.R
\name{sortPairs}
\alias{sortPairs}
\title{Sort Pairs}
\usage{
sortPairs(c1, c2, spMat = FALSE)
}
\description{
`r lifecycle::badge("deprecated")`

A function to sort pairs of integers or factors and identify the pairs between two classifications

Just a change in the function name: please use [sort_pairs()].
}
\examples{
data(iris)
cl <- cutree(hclust(dist(iris[, -5])), 4)
out <- sortPairs(cl, iris$Species)
# ->
out <- sort_pairs(cl, iris$Species)
}
\keyword{internal}
