% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aricode.R
\name{sortPairs}
\alias{sortPairs}
\title{Sort Pairs}
\usage{
sortPairs(c1, c2, spMat = FALSE)
}
\arguments{
\item{c1}{a vector of length n with value between 0 and N1 < n}

\item{c2}{a vector of integer of length n with value between 0 and N2 < n}

\item{spMat}{logical: send back the contigency table a sparsely encoded (cost more than the algorithm itself). Default is FALSE}
}
\description{
A function to sort pairs of integers or factors and identify the pairs
}
